Refactor metadata into own module
